backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pypdf>=3.17.0
pydantic>=2.5.0
aiofiles>=23.2.1
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
PyPDF2>=4.0.0
python-docx>=0.8.11
